Change #1
This commit is extremely long, so I will name the things that I have
completed so far; there are still some changes the need to be made:
-I have added a background with an array list of 100 lines that keep
moving on the screen, and in order to keep it moving: I made it so that
the direction is multiplying by the inputted speed constantly.
-I have also added a circle that is supposed to move when the space key
(or any key) is pressed. That is why an if key pressed is added.
-As well as that, I added a boolean, which essentially ensures that if
a collision between the ball and any line is to happen… then the game
will restart.